release version

+++ app/internal/domain/entities/bond.go
@@ -4,7 +4,7 @@ type Bond struct {
 	Id       *uint16
 	Name     string
 	Nominal  uint16
-	Coupon   uint16
+	Coupon   float32
 	Months   string
 	Day      uint16
 	Quantity uint32
